requirements.txt
asgiref==3.2.10
colorama==0.4.3
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
